move more templates around
