+++ src/templates/inline/index.js
@@ -1,5 +1,4 @@
-const keyValue = require('../parsers/key-value');
-const getInside = require('../parsers/inside');
+const keyValue = require('../parsers/keyValue');
 const pipeSplit = require('../parsers/pipeSplit');
 const strip = require('../parsers/_strip');
 
@@ -84,63 +83,6 @@ const inline = {
     return result.join(' ');
   },
 
-  //a convulated way to make a xml tag - https://en.wikipedia.org/wiki/Template:Tag
-  tag: (tmpl) => {
-    let obj = keyValue(tmpl);
-    if (obj.content) {
-      let order = ['tagName', 'open'];
-      let tagName = pipeSplit(tmpl, order).tagName;
-      //ignore ref tags and all that
-      if (tagName !== 'span' && tagName !== 'div') {
-        return '';
-      }
-      return obj.content.text();
-    }
-    return '';
-  },
-  plural: (tmpl) => {
-    let order = ['num', 'word'];
-    let obj = pipeSplit(tmpl, order);
-    let num = Number(obj.num);
-    let word = obj.word;
-    if (num !== 1) {
-      if (/.y$/.test(word)) {
-        word = word.replace(/y$/, 'ies');
-      } else {
-        word += 's';
-      }
-    }
-    return num + ' ' + word;
-  },
-  'first word': (tmpl) => {
-    let str = getInside(tmpl).data || '';
-    return str.split(' ')[0];
-  },
-  'trunc': (tmpl) => {
-    let order = ['str', 'len'];
-    let obj = pipeSplit(tmpl, order);
-    return obj.str.substr(0, obj.len);
-  },
-  'str mid': (tmpl) => {
-    let order = ['str', 'start', 'end'];
-    let obj = pipeSplit(tmpl, order);
-    let start = parseInt(obj.start, 10) - 1;
-    let end = parseInt(obj.end, 10);
-    return obj.str.substr(start, end);
-  },
-  //grab the first, second or third pipe
-  'p1': (tmpl) => {
-    let order = ['one'];
-    return pipeSplit(tmpl, order).one;
-  },
-  'p2': (tmpl) => {
-    let order = ['one', 'two'];
-    return pipeSplit(tmpl, order).two;
-  },
-  'p3': (tmpl) => {
-    let order = ['one', 'two', 'three'];
-    return pipeSplit(tmpl, order).three;
-  },
   'block indent': (tmpl) => {
     let obj = keyValue(tmpl);
     if (obj['1']) {
